+++ Epoch.net/LongEpochTimeValueException.cs
@@ -4,7 +4,7 @@ namespace Epoch.net
 {
     public class LongEpochTimeValueException : Exception
     {
-        private const string TimeSpanErrorMessage = "";
+        private const string TimeSpanErrorMessage = "The provided TimeSpan {0} does not conform to the LongEpochTime range";
         
         public LongEpochTimeValueException(TimeSpan value)
             :base(string.Format(TimeSpanErrorMessage, value))